+++ CHANGELOG.md
@@ -7,10 +7,11 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 <!-- references -->
 
-[Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
-[Semantic Versioning]: https://semver.org/spec/v2.0.0.html
+[keep a changelog]: https://keepachangelog.com/en/1.0.0/
+[semantic versioning]: https://semver.org/spec/v2.0.0.html
+[bc]: https://github.com/dogmatiq/.github/blob/main/VERSIONING.md#changelogs
 
-## [Unreleased]
+## [0.2.3] - 2025-10-28
 
 ### Added
 
@@ -74,6 +75,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.2.0]: https://github.com/dogmatiq/spruce/releases/tag/v0.2.0
 [0.2.1]: https://github.com/dogmatiq/spruce/releases/tag/v0.2.1
 [0.2.2]: https://github.com/dogmatiq/spruce/releases/tag/v0.2.2
+[0.2.3]: https://github.com/dogmatiq/spruce/releases/tag/v0.2.3
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD